refactor based on feedbacks
- extract MoveHistory class
- rename package
