+++ src/main/kotlin/com/nextstep/racingcar/domain/Car.kt
@@ -1,20 +1,16 @@
 package com.nextstep.racingcar.domain
 
-import com.nextstep.racingcar.domain.gamerules.MoveRule
+import com.nextstep.racingcar.domain.rules.MoveRule
 
 class Car {
-    private val histories = mutableListOf<Movement>()
+    private val histories = MoveHistory()
 
     fun move(moveRule: MoveRule) {
         val movement = moveRule.move()
-        saveHistory(movement)
-    }
-
-    private fun saveHistory(movement: Movement) {
-        histories.add(movement)
+        histories.saveHistory(movement)
     }
 
     fun getHistories(): List<Movement> {
-        return histories.toList()
+        return histories.getHistories()
     }
 }
