+++ src/main/kotlin/com/nextstep/racingcar/domain/MoveHistory.kt
@@ -1,0 +1,13 @@
+package com.nextstep.racingcar.domain
+
+class MoveHistory {
+    private val histories = mutableListOf<Movement>()
+
+    fun saveHistory(movement: Movement) {
+        histories.add(movement)
+    }
+
+    fun getHistories(): List<Movement> {
+        return histories.toList()
+    }
+}
