+++ src/main/kotlin/com/nextstep/racingcar/domain/rules/RandomNumberGenerator.kt
@@ -1,4 +1,4 @@
-package com.nextstep.racingcar.domain.gamerules
+package com.nextstep.racingcar.domain.rules
 
 class RandomNumberGenerator : NumberGenerator {
